backend/pyproject.toml
[project]
name = "kanban-backend"
version = "0.1.0"
description = "Minimal FastAPI backend for MVP skeleton"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.118.0",
  "httpx>=0.28.0",
  "uvicorn[standard]>=0.37.0",
  "pyjwt>=2.9.0",
  "bcrypt>=4.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
